+++ src/components/index.js
@@ -3,4 +3,4 @@ export { default as Card } from './Card';
 export { default as ModalContainer } from './Modal';
 export { default as Navigation } from './Navigation';
 export { default as NotificationContainer } from './Notification';
-export { default as SampleContent } from './SampleContent';
+export { default as Content } from './Content';