+++ src/pages/About/About.js
@@ -1,10 +1,9 @@
 import React from 'react';
-import SampleContent from 'components/SampleContent';
+import Content from 'components/Content';
 
 const AboutPage = () => (
   <div>
-    <h2>About Us</h2>
-		<SampleContent numParagraphs={3} paragraphLength={5} />
+		<Content title="About Us" />
   </div>
 );