+++ src/pages/Cards/Cards.js
@@ -1,13 +1,12 @@
 import React from 'react';
 import Card from 'components/Card';
-import SampleContent from 'components/SampleContent';
+import Content from 'components/Content';
 import profiles from 'data/profiles.json';
 import './Cards.scss';
 
 const CardsPage = () => (
 	<div>
-		<h2>Cards</h2>
-		<SampleContent paragraphLength={5} />
+		<Content title="Cards" sampleLength="3" />
 		<div className="card-container">
 			{profiles.map((profile, index) => <Card key={index} {...profile} />)}
 		</div>